+++ CHANGES.md
@@ -4,6 +4,9 @@
 
 - Support for OCaml 5.0
 
+- New functions `Py.Err.set_interrupt` and, for Python >=3.10,
+ `Py.Err.set_interrupt_ex`.
+
 - New functions
   `Py.Dict.{to_bindings_seq, to_bindings_seq_map, to_bindings_string_seq}`.